+++ examples/material/main.rs
@@ -1,9 +1,10 @@
 //! Displays spheres with physically based materials.
 
 extern crate amethyst;
+extern crate amethyst_assets;
 
 use amethyst::{
-    assets::Loader,
+    assets::AssetLoaderSystemData,
     core::{
         cgmath::{Deg, Vector3},
         Transform, TransformBundle,
@@ -22,16 +23,15 @@ impl<'a, 'b> SimpleState<'a, 'b> for Example {
 
         println!("Load mesh");
         let (mesh, albedo) = {
-            let loader = world.read_resource::<Loader>();
-
-            let meshes = &world.read_resource();
-            let textures = &world.read_resource();
-            let mesh: MeshHandle = loader.load_from_data(
-                Shape::Sphere(32, 32).generate::<Vec<PosNormTangTex>>(None),
-                (),
-                meshes,
-            );
-            let albedo = loader.load_from_data([1.0, 1.0, 1.0, 1.0].into(), (), textures);
+            let mesh = world.exec(|loader: AssetLoaderSystemData<Mesh>| {
+                loader.load_from_data(
+                    Shape::Sphere(32, 32).generate::<Vec<PosNormTangTex>>(None),
+                    (),
+                )
+            });
+            let albedo = world.exec(|loader: AssetLoaderSystemData<Texture>| {
+                loader.load_from_data([1.0, 1.0, 1.0, 1.0].into(), ())
+            });
 
             (mesh, albedo)
         };
@@ -52,15 +52,12 @@ impl<'a, 'b> SimpleState<'a, 'b> for Example {
                 let metallic = [metallic, metallic, metallic, 1.0].into();
                 let roughness = [roughness, roughness, roughness, 1.0].into();
 
-                let (metallic, roughness) = {
-                    let loader = world.read_resource::<Loader>();
-                    let textures = &world.read_resource();
-
-                    let metallic = loader.load_from_data(metallic, (), textures);
-                    let roughness = loader.load_from_data(roughness, (), textures);
-
-                    (metallic, roughness)
-                };
+                let (metallic, roughness) = world.exec(|loader: AssetLoaderSystemData<Texture>| {
+                    (
+                        loader.load_from_data(metallic, ()),
+                        loader.load_from_data(roughness, ()),
+                    )
+                });
 
                 let mtl = Material {
                     albedo: albedo.clone(),